requirements.txt
django
djangorestframework
python_dotenv
qrcode
Pillow
mysqlclient
drf-yasg
